시작 스테이지 설계
시작 스테이지의 대략적인 뼈대를 csv파일로 만들어 보았다. 시작 스테이지의 관리와 구체적인 기능 구현은 브렌치를 합친 후 구현해야 할 듯 하다.

+++ RogueLike_Project/Assets/Scripts/TileManager.cs
@@ -10,13 +10,13 @@ public class TileManager : MonoBehaviour
     public static int mapSize = 30;
     Tile[,] tiles = new Tile[mapSize, mapSize];
     [SerializeField] GameObject tile;
-    int[,] tileMap = new int[mapSize, mapSize];
+    float[,] tileMap = new float[mapSize, mapSize];
     CSVToArray CTA;
     public int GetMapSize
     {
         get { return mapSize; }
     }
-    public int[,] GetTileMap
+    public float[,] GetTileMap
     { 
         get { return tileMap; }
     }
